zilany2009: fix indexing with floats in ffGn

+++ cochlea/zilany2009/ffGn_module.py
@@ -21,9 +21,9 @@ def ffGn(N, tdres, Hinput, mu):
 
 
     # Downsampling No. of points to match with those of Scott jackson (tau 1e-1)
-    resamp = np.ceil(1e-1 / tdres)
+    resamp = int(np.ceil(1e-1 / tdres))
     nop = N
-    N = np.ceil(N / resamp) + 1
+    N = int(np.ceil(N / resamp) + 1)
     if N < 10:
         N = 10
 
@@ -43,7 +43,7 @@ def ffGn(N, tdres, Hinput, mu):
         y = randn(N)
     else:
         # TODO: make variables persistant
-        Nfft = 2 ** np.ceil(np.log2(2*(N-1)))
+        Nfft = int(2 ** np.ceil(np.log2(2*(N-1))))
         NfftHalf = np.round(Nfft / 2)
 
         k = np.concatenate( (np.arange(0,NfftHalf), np.arange(NfftHalf,0,-1)) )
